Added main-splash page, Also added mickyBack functionaility
