+++ js/listeners.js
@@ -10,6 +10,7 @@ let cindys = require('./cindys');
 let frontierLand = require("./frontierLand");
 let mstObj = require("./main");
 let modalWindow = require("../templates/modal-window.hbs");
+let mainTemplate = require("../templates/master-template.hbs");
 // let mainOb = require("./main");
 
 // LOCATION LISTENERS
